@@ -9,32 +9,34 @@
   "bugs": {
     "url": "https://github.com/graememcc/micropolisJS/issues"
   },
+  "type": "module",
   "scripts": {
-    "build": "webpack --env.production",
-    "watch": "webpack --watch --env.development",
+    "build": "webpack --mode production",
+    "watch": "webpack --watch --mode development",
     "test": "jest",
     "test:watch": "jest --watch",
     "lint": "tslint src/*.ts test/*.ts",
-    "dev": "webpack-dev-server --env.development",
-    "production": "webpack-dev-server --env.production"
+    "dev": "webpack serve --mode development",
+    "production": "webpack serve --mode production"
   },
   "devDependencies": {
-    "@types/jest": "23.1.0",
-    "awesome-typescript-loader": "5.0.0",
-    "clean-webpack-plugin": "0.1.19",
-    "copy-webpack-plugin": "^4.5.1",
-    "git-revision-webpack-plugin": "3.0.1",
-    "html-webpack-plugin": "3.2.0",
-    "jest": "23.1.0",
-    "script-ext-html-webpack-plugin": "2.0.1",
-    "ts-jest": "22.4.4",
-    "tslint": "5.10.0",
-    "typescript": "2.9.2",
-    "webpack": "4.6.0",
-    "webpack-cli": "2.0.15",
-    "webpack-dev-server": "3.1.14"
+    "@types/jest": "30.0.0",
+    "clean-webpack-plugin": "4.0.0",
+    "copy-webpack-plugin": "13.0.0",
+    "git-revision-webpack-plugin": "5.0.0",
+    "html-webpack-plugin": "5.6.3",
+    "jest": "30.0.4",
+    "ts-jest": "29.4.0",
+    "ts-loader": "9.5.2",
+    "typescript": "5.8.3",
+    "webpack": "5.100.1",
+    "webpack-cli": "6.0.1",
+    "webpack-dev-server": "5.2.2"
   },
   "author": "Graeme McCutcheon (http://www.graememcc.co.uk)",
   "license": "GPL-3.0",
-  "readmeFilename": "README.md"
+  "readmeFilename": "README.md",
+  "dependencies": {
+    "jquery": "3.7.1"
+  }
 }
