@@ -7,8 +7,8 @@
  *
  */
 
-import { BaseTool } from './baseTool';
-import { Connector } from './connector';
+import { BaseTool } from './baseTool.js';
+import { Connector } from './connector.js';
 
 // Take a tool constructor, make it inherit from BaseTool, and add
 // the various connection related functions
