@@ -7,12 +7,12 @@
  *
  */
 
-import { forEachCardinalDirection, getRandomCardinalDirection, getRandomDirection } from './direction';
-import { GameMap } from './gameMap';
-import { Position } from './position';
-import { Random } from './random';
-import { BLBNBIT, BULLBIT } from "./tileFlags";
-import { CHANNEL, DIRT, REDGE, RIVER, WATER_LOW, WATER_HIGH, WOODS, WOODS_LOW, WOODS_HIGH } from "./tileValues";
+import { forEachCardinalDirection, getRandomCardinalDirection, getRandomDirection } from './direction.ts';
+import { GameMap } from './gameMap.js';
+import { Position } from './position.ts';
+import { Random } from './random.ts';
+import { BLBNBIT, BULLBIT } from "./tileFlags.ts";
+import { CHANNEL, DIRT, REDGE, RIVER, WATER_LOW, WATER_HIGH, WOODS, WOODS_LOW, WOODS_HIGH } from "./tileValues.ts";
 
 var TERRAIN_CREATE_ISLAND;
 var TERRAIN_TREE_LEVEL = -1;
