@@ -6,9 +6,10 @@
  * http://micropolisjs.graememcc.co.uk/COPYING
  *
  */
+import $ from "jquery";
 
-import { NAG_WINDOW_CLOSED } from './messages';
-import { ModalWindow } from './modalWindow';
+import { NAG_WINDOW_CLOSED } from './messages.ts';
+import { ModalWindow } from './modalWindow.js';
 
 var NagWindow = ModalWindow(function() {
   $(nagFormID).on('submit', submit.bind(this));
