@@ -6,11 +6,12 @@
  * http://micropolisjs.graememcc.co.uk/COPYING
  *
  */
+import $ from "jquery";
 
-import { Config } from './config';
-import { QUERY_WINDOW_CLOSED } from './messages';
-import { ModalWindow } from './modalWindow';
-import { MiscUtils } from './miscUtils';
+import { Config } from './config.js';
+import { QUERY_WINDOW_CLOSED } from './messages.ts';
+import { ModalWindow } from './modalWindow.js';
+import { MiscUtils } from './miscUtils.js';
 
 var QueryWindow = ModalWindow(function() {
   this._debugToggled = false;
