@@ -7,11 +7,11 @@
  *
  */
 
-import { Random } from './random';
-import { Tile } from "./tile";
-import { ANIMBIT, BIT_MASK, BULLBIT } from "./tileFlags";
-import * as TileValues from "./tileValues";
-import { TileUtils } from './tileUtils';
+import { Random } from './random.ts';
+import { Tile } from "./tile.ts";
+import { ANIMBIT, BIT_MASK, BULLBIT } from "./tileFlags.ts";
+import * as TileValues from "./tileValues.ts";
+import { TileUtils } from './tileUtils.js';
 
 var openBridge = function(map, origX, origY, xDelta, yDelta, oldTiles, newTiles) {
   for (var i = 0; i < 7; i++) {
