@@ -13,4 +13,4 @@ export const SPRITE_AIRPLANE = 3;
 export const SPRITE_SHIP = 4;
 export const SPRITE_MONSTER = 5;
 export const SPRITE_TORNADO = 6;
-export const SPRTE_EXPLOSION = 7;
+export const SPRITE_EXPLOSION = 7;
